feat: create login page, and more maintenale code
